rgw: fix (squash)
Call FCGX_Finish_r() on fcgx when ending request. This will not close
the connection if module requested to keep connection open. This is not
really necessary, as the following call on FCGX_Accept_r() would do the
same, but will be needed if size of queue ring is larger than the number
of connections.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>